GP-6176: More Objective-C type metadata analyzers refactoring
